src/notification_service/requirements.txt
Flask==2.2.2
google-cloud-secret-manager==2.16.2
